AMD的温度和NVIDA类型不一样，一个是uint一个是int

+++ src/NTMiner.Core/Core/Gpus/Impl/NVIDIAGpuSet.cs
@@ -130,7 +130,7 @@ namespace NTMiner.Core.Gpus.Impl {
 
                     Gpu gpu = (Gpu)_gpus[i];
                     bool isChanged = gpu.Temperature != temp || gpu.PowerUsage != power || gpu.FanSpeed != speed;
-                    gpu.Temperature = temp;
+                    gpu.Temperature = (int)temp;
                     gpu.PowerUsage = power;
                     gpu.FanSpeed = speed;
